refactor: extract chunk producer logic from Client (#12852)
Reducing number of fields in Client from 36 to 31 and number of lines in
client.rs from 2913 to 2624.
This is done by separating logic exclusive to chunk production, thus new
struct is named `ChunkProducer`. It is almost as simple as moving some
methods from Client to ChunkProducer.

I had to pass `chain_validate: &dyn Fn(&SignedTransaction) -> bool,`
through which is not great but already used in other places. Also, I
remove `mut` keywords because the closure doesn't mutate neither
environment nor variable.

Later, we can also hide `produce_chunks` and state witness production
logic. This process is a bit heavy because currently access to the whole
Client is used everywhere - which is not really needed.
Another next step can be separating BlockProducer logic.
