+++ chain/chain/src/stateless_validation/chunk_validation.rs
@@ -389,7 +389,7 @@ pub fn pre_validate_chunk_state_witness(
             } else {
                 let prev_block_header =
                     store.get_block_header(last_chunk_block.header().prev_hash())?;
-                let mut check = chain.transaction_validity_check(prev_block_header);
+                let check = chain.transaction_validity_check(prev_block_header);
                 state_witness.transactions.iter().map(|t| check(t)).collect::<Vec<_>>()
             }
         } else {
