+++ integration-tests/src/tests/client/features/adversarial_behaviors.rs
@@ -366,7 +366,7 @@ fn test_banning_chunk_producer_when_seeing_invalid_chunk_base(
 fn slow_test_banning_chunk_producer_when_seeing_invalid_chunk() {
     init_test_logger();
     let mut test = AdversarialBehaviorTestData::new();
-    test.env.clients[7].produce_invalid_chunks = true;
+    test.env.clients[7].chunk_producer.produce_invalid_chunks = true;
     test_banning_chunk_producer_when_seeing_invalid_chunk_base(test);
 }
 
@@ -378,7 +378,7 @@ fn test_banning_chunk_producer_when_seeing_invalid_tx_in_chunk() {
     if !relaxed_chunk_validation {
         init_test_logger();
         let mut test = AdversarialBehaviorTestData::new();
-        test.env.clients[7].produce_invalid_tx_in_chunks = true;
+        test.env.clients[7].chunk_producer.produce_invalid_tx_in_chunks = true;
         test_banning_chunk_producer_when_seeing_invalid_chunk_base(test);
     }
     // Otherwise the chunks aren't considered invalid and there will be no banning.
